backend/requirements.txt
# Kizuna Backend - Python/FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
supabase==2.10.0
google-generativeai==0.8.0
python-telegram-bot==21.5
httpx==0.27.0
pydantic==2.9.0
python-multipart==0.0.9
